Part 4 Backend Finished

+++ part4/src/routes/blogs.routes.js
@@ -1,9 +1,11 @@
 import { Router } from "express";
 import { Blog } from "../../schemas/blogSchema.js";
+import verifyLoggedIn from "../../middlewares/jsonToken.js";
 const router = Router();
 
 router.get("/blogs", (req, res) => {
   Blog.find()
+    .populate("author")
     .then((blogs) => {
       console.log(blogs);
       res.status(200).json({
@@ -14,9 +16,9 @@ router.get("/blogs", (req, res) => {
     .catch((error) => res.status(400).json({ errorMessage: error.message }));
 });
 
-router.post("/blogs/create", (req, res) => {
-  const { title, author, url, likes } = req.body;
-  Blog.create({ title, author, url, likes })
+router.post("/blogs/create", verifyLoggedIn, (req, res) => {
+  const { title, url, likes } = req.body;
+  Blog.create({ title, author: req.userId, url, likes })
     .then((blog) => {
       console.log(blog);
       res.status(201).json({
@@ -27,11 +29,11 @@ router.post("/blogs/create", (req, res) => {
     .catch((error) => res.status(400).json({ errorMessage: error.message }));
 });
 
-router.put("/blogs/update/:id", (req, res) => {
+router.put("/blogs/update/:id", verifyLoggedIn, (req, res) => {
   const { id } = req.params;
   const { likes } = req.body;
   Blog.updateOne(
-    { _id: id },
+    { _id: id, author: req.userId },
     {
       likes,
     }
@@ -46,8 +48,8 @@ router.put("/blogs/update/:id", (req, res) => {
     .catch((error) => res.status(400).json({ errorMessage: error.message }));
 });
 
-router.delete("/blogs/delete/:id", (req, res) => {
-  Blog.deleteOne({ _id: req.params.id })
+router.delete("/blogs/delete/:id", verifyLoggedIn, (req, res) => {
+  Blog.deleteOne({ _id: req.params.id, author: req.userId })
     .then((blog) => {
       console.log(blog);
       res.status(200).json({
